requirements.txt
discord.py
pyTigerGraph
requests
nltk
numpy